@@ -3,31 +3,28 @@
     "description": "Swoole script for BEAR.Sunday",
     "type": "library",
     "require": {
-        "php": "^7.3 || ^8.0",
-        "ext-json": "*",
+        "php": "^8.2",
         "ext-swoole": ">=4.0.0",
         "bear/app-meta": "^1.6",
         "bear/query-repository": "^1.8.4",
-        "bear/resource": "^1.15",
-        "bear/sunday": "^1.5",
-        "doctrine/cache": "^1.10",
+        "bear/resource": "^1.17",
+        "bear/sunday": "^1.6",
         "fastd/http": "5.0.x-dev",
         "psr/cache": "^1.0 || ^2.0 || ^3.0",
-        "psr/http-message": "^1.0",
-        "ray/di": "^2.11.3",
+        "psr/http-message": "^1.0 || ^2.0",
+        "ray/di": "^2.14",
         "ray/psr-cache-module": "^1.3",
         "ray/psr7-module": "^1.3",
-        "symfony/cache": "^5.4 || ^6.2"
+        "symfony/cache": "^5.4 || ^6.0 || ^7.0"
     },
     "require-dev": {
         "bamarni/composer-bin-plugin": "^1.4",
-        "bear/package": "^1.9",
-        "bear/skeleton": "^1.6",
-        "eaglewu/swoole-ide-helper": "dev-master",
-        "guzzlehttp/guzzle": "^6.5 || ^7.5",
-        "josegonzalez/dotenv": "^3.2",
-        "phpunit/phpunit": "^9.5",
-        "symfony/process": "^5.0"
+        "bear/package": "^1.14",
+        "bear/skeleton": "^1.10",
+        "guzzlehttp/guzzle": "^7.5",
+        "phpunit/phpunit": "^10.0 || ^11.0",
+        "roave/security-advisories": "dev-latest",
+        "symfony/process": "^5.0 || ^6.0 || ^7.0"
     },
     "license": "MIT",
     "authors": [
