@@ -7,22 +7,15 @@ namespace BEAR\Swoole;
 use BEAR\QueryRepository\ResourceStorageInterface;
 use Swoole\Http\Response;
 
-final class HttpCache implements HttpCacheInterface
+final readonly class HttpCache implements HttpCacheInterface
 {
-    /**
-     * @var ResourceStorageInterface
-     */
-    private $storage;
-
-    public function __construct(ResourceStorageInterface $storage)
-    {
-        $this->storage = $storage;
+    public function __construct(
+        private ResourceStorageInterface $storage,
+    ) {
     }
 
-    /**
-     * @param array<string, string> $server
-     */
-    public function isNotModified(array $server) : bool
+    /** @param array<string, string> $server */
+    public function isNotModified(array $server): bool
     {
         return isset($server['if-none-match']) && $this->storage->hasEtag($server['if-none-match']);
     }
