@@ -10,18 +10,16 @@ use Swoole\Http\Response;
 
 final class Responder implements TransferInterface
 {
-    /**
-     * @var Response
-     */
-    private $response;
+    private Response $response;
 
-    public function __invoke(ResourceObject $ro, array $server)
+    public function __invoke(ResourceObject $ro, array $server): void
     {
         unset($server);
         $ro->toString();
         foreach ($ro->headers as $key => $value) {
             $this->response->header($key, (string) $value);
         }
+
         $this->response->status($ro->code);
         $this->response->end($ro->view);
     }
