@@ -7,53 +7,61 @@
  * Original author: jan huang <bboyjanhuang@gmail.com>
  */
 
+declare(strict_types=1);
+
 namespace BEAR\Swoole;
 
 use FastD\Http\ServerRequest;
 use Swoole\Http\Request;
 
-/**
- * @SuppressWarnings
- */
+use function explode;
+use function parse_url;
+use function str_replace;
+
+use const PHP_URL_HOST;
+use const SWOOLE_VERSION;
+
+/** @SuppressWarnings(PHPMD) */
 class SwooleServerRequest extends ServerRequest
 {
     public static function createServerRequestFromSwoole(Request $request): ServerRequest
     {
-        $get = isset($request->get) ? $request->get : [];
-        $post = isset($request->post) ? $request->post : [];
-        $cookie = isset($request->cookie) ? $request->cookie : [];
-        $files = isset($request->files) ? $request->files : [];
+        $get = $request->get ?? [];
+        $post = $request->post ?? [];
+        $cookie = $request->cookie ?? [];
+        $files = $request->files ?? [];
 
         $host = '::1';
         foreach (['host', 'server_addr'] as $name) {
-            if (!empty($request->header[$name])) {
+            if (! empty($request->header[$name])) {
                 $host = parse_url($request->header[$name], PHP_URL_HOST) ?: $request->header[$name];
             }
         }
 
+        $serverProtocol = $request->server['server_protocol'] ?? 'HTTP/1.1';
         $server = [
             'REQUEST_METHOD' => $request->server['request_method'],
             'REQUEST_URI' => $request->server['request_uri'],
-            'PATH_INFO' => $request->server['path_info'],
-            'REQUEST_TIME' => $request->server['request_time'],
+            'PATH_INFO' => $request->server['path_info'] ?? '',
+            'REQUEST_TIME' => $request->server['request_time'] ?? time(),
             'GATEWAY_INTERFACE' => 'swoole/' . SWOOLE_VERSION,
             // Server
-            'SERVER_PROTOCOL' => isset($request->header['server_protocol']) ? $request->header['server_protocol'] : $request->server['server_protocol'],
-            'REQUEST_SCHEMA' => isset($request->header['request_scheme']) ? $request->header['request_scheme'] : explode('/', $request->server['server_protocol'])[0],
-            'SERVER_NAME' => isset($request->header['server_name']) ? $request->header['server_name'] : $host,
+            'SERVER_PROTOCOL' => $request->header['server_protocol'] ?? $serverProtocol,
+            'REQUEST_SCHEMA' => $request->header['request_scheme'] ?? explode('/', $serverProtocol)[0],
+            'SERVER_NAME' => $request->header['server_name'] ?? $host,
             'SERVER_ADDR' => $host,
-            'SERVER_PORT' => isset($request->header['server_port']) ? $request->header['server_port'] : $request->server['server_port'],
-            'REMOTE_ADDR' => $request->server['remote_addr'],
-            'REMOTE_PORT' => isset($request->header['remote_port']) ? $request->header['remote_port'] : $request->server['remote_port'],
-            'QUERY_STRING' => isset($request->server['query_string']) ? $request->server['query_string'] : '',
+            'SERVER_PORT' => $request->header['server_port'] ?? $request->server['server_port'] ?? 80,
+            'REMOTE_ADDR' => $request->server['remote_addr'] ?? '127.0.0.1',
+            'REMOTE_PORT' => $request->header['remote_port'] ?? $request->server['remote_port'] ?? 0,
+            'QUERY_STRING' => $request->server['query_string'] ?? '',
             // Headers
             'HTTP_HOST' => $host,
-            'HTTP_USER_AGENT' => isset($request->header['user-agent']) ? $request->header['user-agent'] : '',
-            'HTTP_ACCEPT' => isset($request->header['accept']) ? $request->header['accept'] : '*/*',
-            'HTTP_ACCEPT_LANGUAGE' => isset($request->header['accept-language']) ? $request->header['accept-language'] : '',
-            'HTTP_ACCEPT_ENCODING' => isset($request->header['accept-encoding']) ? $request->header['accept-encoding'] : '',
-            'HTTP_CONNECTION' => isset($request->header['connection']) ? $request->header['connection'] : '',
-            'HTTP_CACHE_CONTROL' => isset($request->header['cache-control']) ? $request->header['cache-control'] : '',
+            'HTTP_USER_AGENT' => $request->header['user-agent'] ?? '',
+            'HTTP_ACCEPT' => $request->header['accept'] ?? '*/*',
+            'HTTP_ACCEPT_LANGUAGE' => $request->header['accept-language'] ?? '',
+            'HTTP_ACCEPT_ENCODING' => $request->header['accept-encoding'] ?? '',
+            'HTTP_CONNECTION' => $request->header['connection'] ?? '',
+            'HTTP_CACHE_CONTROL' => $request->header['cache-control'] ?? '',
         ];
 
         $headers = [];
@@ -66,7 +74,7 @@ class SwooleServerRequest extends ServerRequest
             static::createUriFromGlobal($server),
             $headers,
             null,
-            $server
+            $server,
         );
         unset($headers);
 
@@ -76,7 +84,6 @@ class SwooleServerRequest extends ServerRequest
             ->withParsedBody($post)
             ->withQueryParams($get)
             ->withCookieParams($cookie)
-            ->withUploadedFiles($files)
-            ;
+            ->withUploadedFiles($files);
     }
 }
