@@ -4,49 +4,45 @@ declare(strict_types=1);
 
 namespace BEAR\Swoole;
 
+use RuntimeException;
+use Symfony\Component\Process\Process;
+
 use function error_log;
 use function register_shutdown_function;
-use RuntimeException;
 use function sprintf;
-use function strpos;
-use Symfony\Component\Process\Process;
-use function var_dump;
+use function str_contains;
+
+use const PHP_BINARY;
 
 final class SwooleServer
 {
-    /**
-     * @var Process<mixed>
-     */
-    private $process;
+    /** @var Process<mixed> */
+    private Process $process;
 
     public function __construct(string $phpFile)
     {
         $this->process = new Process([
             PHP_BINARY,
-            $phpFile
+            $phpFile,
         ]);
-        register_shutdown_function(function () {
+        register_shutdown_function(function (): void {
             $this->process->stop();
         });
     }
 
-    public function start() : void
+    public function start(): void
     {
         $this->process->start();
-        if (version_compare(PHP_VERSION, '7.4.0', '<')) {
-            sleep(1);
-
-            return;
-        }
-        $this->process->waitUntil(function (string $type, string $output) : bool {
+        $this->process->waitUntil(function (string $type, string $output): bool {
             if ($type === 'err') {
                 error_log($output);
             }
-            return (bool) strpos($output, 'started');
+
+            return str_contains($output, 'started');
         });
     }
 
-    public function stop() : void
+    public function stop(): void
     {
         $exitCode = $this->process->stop();
         if ($exitCode !== 143) {
