@@ -9,7 +9,7 @@ use Ray\HttpMessage\RequestProviderInterface;
 
 final class SwooleRequestProvider implements RequestProviderInterface
 {
-    public function get() : ServerRequestInterface
+    public function get(): ServerRequestInterface
     {
         return SwooleServerRequest::createServerRequestFromSwoole(SuperGlobals::$swooleRequest);
     }
