@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace BEAR\Skeleton\Resource\Page;
 
 use BEAR\Resource\ResourceObject;
@@ -9,19 +12,17 @@ use Ray\WebContextParam\Annotation\ServerParam;
 
 class WebContext extends ResourceObject
 {
-    /**
-     * @CookieParam(param="cookie", key="c")
-     * @FormParam(param="form", key="f")
-     * @QueryParam(param="query", key="q")
-     * @ServerParam(param="header", key="HTTP_X_MY_HEADER")
-     */
-    public function onPost(string $cookie, string $form, string $query, string $header): ResourceObject
-    {
+    public function onPost(
+        #[CookieParam('c')] string $cookie,
+        #[FormParam('f')] string $form,
+        #[QueryParam('q')] string $query,
+        #[ServerParam('HTTP_X_MY_HEADER')] string $header,
+    ): static {
         $this->body = [
             'cookie' => $cookie,
             'form' => $form,
             'query' => $query,
-            'header' => $header
+            'header' => $header,
         ];
 
         return $this;
